--- frontend/src/types/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-  id: string;
-  username: string;
-  joined_at: string;
-}
